+++ examples/simple_verbose.rs
@@ -1,8 +1,8 @@
 #[allow(unused_imports)]
-extern crate failure;
-extern crate structopt_flags;
 #[macro_use]
 extern crate structopt;
+extern crate failure;
+extern crate structopt_flags;
 
 use failure::Error;
 use structopt::StructOpt;
@@ -10,7 +10,7 @@ use structopt::StructOpt;
 #[derive(Debug, StructOpt)]
 #[structopt(
     name = "simple_verbose",
-    about = "An example using verbose flag"
+    about = "An example using simple verbose flag"
 )]
 struct Opt {
     #[structopt(flatten)]