+++ src/ipv4.rs
@@ -1,4 +1,4 @@
-use super::GetOrDefault;
+use super::GetWithDefault;
 use super::Ipv4Address;
 use std::fmt;
 use std::net::Ipv4Addr;
@@ -12,8 +12,9 @@ use std::net::Ipv4Addr;
 /// #[macro_use]
 /// extern crate structopt;
 ///
+/// use std::net::Ipv4Addr;
 /// use structopt::StructOpt;
-/// use structopt_flags::GetOrDefault; // to access get_ipv4_addr
+/// use structopt_flags::GetWithDefault; // to access get_ipv4_addr
 ///
 /// #[derive(Debug, StructOpt)]
 /// #[structopt(name = "log_level_opt", about = "An example using the LogLevelOpt option")]
@@ -24,8 +25,7 @@ use std::net::Ipv4Addr;
 ///
 /// fn main() {
 ///     let opt = Opt::from_args();
-///     let ipv4_default = "127.0.0.1".parse().unwrap();
-///     let ipv4 = opt.host_ip.get_or_default(ipv4_default);
+///     let ipv4 = opt.host_ip.get_with_default(Ipv4Addr::new(127,0,0,1));
 /// }
 /// ```
 
@@ -57,11 +57,9 @@ impl fmt::Display for HostV4Opt {
     }
 }
 
-impl GetOrDefault<Ipv4Addr> for HostV4Opt {
-    fn get_or_default(&self, default: Ipv4Addr) -> Ipv4Addr {
-        match self.host_addr {
-            Some(x) => x,
-            None => default,
-        }
+impl GetWithDefault for HostV4Opt {
+    type Item = Ipv4Addr;
+    fn get_with_default(&self, default: Self::Item) -> Self::Item {
+        self.host_addr.unwrap_or(default)
     }
 }