+++ src/lib.rs
@@ -31,6 +31,7 @@ mod verbose;
 
 pub use log::{Level, LevelFilter};
 /// This trait is designed to provide a log level compatible with the Log crates
+/// Options or flags that can provide a log level, will implement this trait
 pub trait LogLevel {
     /// Return the log level.
     ///
@@ -43,6 +44,16 @@ pub trait LogLevel {
     fn get_level_filter(&self) -> LevelFilter;
 }
 
+/// This trait is designed to provide a rude form of default value for options
+/// If an option doesn't have a default value, it will implement this trait
+pub trait GetWithDefault {
+    type Item;
+
+    /// This function can be used to retrieve the value of the command line option
+    /// taking in account the default value used as argument
+    fn get_with_default(&self, default: Self::Item) -> Self::Item;
+}
+
 pub use std::net::Ipv4Addr;
 
 pub trait Ipv4Address {
@@ -50,12 +61,11 @@ pub trait Ipv4Address {
 }
 
 pub use ipv4::HostV4Opt;
+pub use logopt::LogLevelNoDef;
+pub use logopt::LogLevelNoDefLower;
 pub use logopt::LogLevelOpt;
 pub use logopt::LogLevelOptLower;
 pub use verbose::QuietVerbose;
 pub use verbose::SimpleVerbose;
 pub use verbose::Verbose;
-
-pub trait GetOrDefault<T> {
-    fn get_or_default(&self, default: T) -> T;
-}
+pub use verbose::VerboseNoDef;