+++ src/verbose.rs
@@ -1,4 +1,4 @@
-use super::LogLevel;
+use super::{GetWithDefault, LogLevel};
 use log::{Level, LevelFilter};
 use std::fmt;
 
@@ -61,6 +61,70 @@ impl LogLevel for Verbose {
 impl fmt::Display for Verbose {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{}", self.get_level_filter())
+    }
+}
+
+/// This struct provides the `--verbose` cli option, without a predefined default
+///
+/// Multiple occurrences, will increase the verbosity level
+///
+/// ```rust
+/// extern crate log;
+/// extern crate structopt_flags;
+/// #[macro_use]
+/// extern crate structopt;
+///
+/// use log::LevelFilter;
+/// use structopt::StructOpt;
+/// use structopt_flags::GetWithDefault; // to access get_log_level
+///
+/// #[derive(Debug, StructOpt)]
+/// #[structopt( name = "verbose_no_def", about = "An example using verbose flag with no predefined default")]
+/// struct Opt {
+///     #[structopt(flatten)]
+///     verbose: structopt_flags::VerboseNoDef,
+/// }
+///
+/// fn main() {
+///     let opt = Opt::from_args();
+///     let filter_level = opt.verbose.get_with_default(LevelFilter::Off);
+///     // set log level
+/// }
+/// ```
+#[derive(StructOpt, Debug, Clone)]
+pub struct VerboseNoDef {
+    /// Increase the output's verbosity level
+    /// Pass many times to increase verbosity level.
+    #[structopt(
+        name = "verbose",
+        long = "verbose",
+        short = "v",
+        parse(from_occurrences),
+        raw(global = "true")
+    )]
+    verbosity_level: u8,
+}
+
+impl GetWithDefault for VerboseNoDef {
+    type Item = LevelFilter;
+    fn get_with_default(&self, default: Self::Item) -> Self::Item {
+        let default_value: i8 = match default {
+            LevelFilter::Off => -1,
+            LevelFilter::Error => 0,
+            LevelFilter::Warn => 1,
+            LevelFilter::Info => 2,
+            LevelFilter::Debug => 3,
+            LevelFilter::Trace => 4,
+        };
+
+        match default_value + self.verbosity_level as i8 {
+            -1 => LevelFilter::Off,
+            0 => LevelFilter::Error,
+            1 => LevelFilter::Warn,
+            2 => LevelFilter::Info,
+            3 => LevelFilter::Debug,
+            _ => LevelFilter::Trace,
+        }
     }
 }